+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 1. [](#new)
     * Added new `onPageHeaders()` event to allow for header modification as needed
     * Added a `system.pages.dirs` configuration option to allow for configurable paths, and multiple page paths
+    * Added new `Pages::getSimplePagesHash` which is useful for caching pages specific data
 1. [](#bugfix)
     * An attempt to workaround windows reading locked file issue [getgrav/grav-plugin-admin#2299](https://github.com/getgrav/grav-plugin-admin/issues/2299)